Fix: Use lfit/checkout-gerrit-change-action
Using actions/checkout is a bug since the workflows are primarily
designed be used for gerrit repositories. The workflows servicing
gerrit repos must use the lfit/checkout-gerrit-change-action for
checkouts. This ensures the refspec is used while checking out the
repo.

Note: A new GERRIT_REFSPEC param has been added to workflow_call
which needs to be updated to downstream workflows or projects.

Signed-off-by: Anil Belur <abelur@linuxfoundation.org>